requirements.txt
opencv-python
pillow
numpy
moviepy
